Merge from develop to master for release v1.0

+++ scripts/burn-image.sh
@@ -6,13 +6,6 @@ if [ -z "$(whoami | grep root)" ]; then
     exit
 fi
 
-working_dir=/tmp/raspbian
-image=https://downloads.raspberrypi.org/raspios_lite_armhf/images/raspios_lite_armhf-2021-11-08/2021-10-30-raspios-bullseye-armhf-lite.zip
-image_hash="008d7377b8c8b853a6663448a3f7688ba98e2805949127a1d9e8859ff96ee1a9"
-archive=raspbian-os-lite.zip
-number_pattern="[0-9]+"
-mount_point=/mnt/temp
-
 setup_environment() {
     if [ ! -d $working_dir ]; then 
         mkdir $working_dir
@@ -25,7 +18,6 @@ cleanup_environment() {
 }
 
 mount_partition () {
-    echo "Mounting partition /dev/$partition."
     if [ ! -d $mount_point ]; then 
         mkdir $mount_point
     fi
@@ -38,27 +30,35 @@ mount_partition () {
     fi
 
     mount "/dev/$partition" $mount_point
-    hdparm -z /dev/$chosen_disk
-    echo "Partition /dev/$partition mounted."
+    hdparm -z /dev/$chosen_disk > /dev/null
 }
 
 unmount_partition () {
-    umount "/dev/$partition"
-    hdparm -z /dev/$chosen_disk
-    if [ -d $mount_point ]; then 
-        rm -rf $mount_point
+    if [ -n "$(mount | grep "$partitions")" ]; then
+        umount "/dev/$partition"
+        hdparm -z /dev/$chosen_disk > /dev/null
+        if [ -d $mount_point ]; then 
+            rm -rf $mount_point
+        fi
+        sleep 3 # Give the OS time to reread
     fi
-    sleep 3 # Give the OS time to reread
-    echo "Partition /dev/$partition unmounted."
 }
+
+working_dir=/tmp/raspbian
+image=https://downloads.raspberrypi.org/raspios_lite_arm64/images/raspios_lite_arm64-2022-01-28/2022-01-28-raspios-bullseye-arm64-lite.zip
+image_hash="d694d2838018cf0d152fe81031dba83182cee79f785c033844b520d222ac12f5"
+archive=2022-01-28-raspios-bullseye-arm64-lite.zip
+mount_point=/mnt/temp
 
 setup_environment
 
 readarray -t disks < <(lsblk -b -e7 -o name,type | grep disk | awk '{print $1}')
 sd_disks=()
 for disk in "${disks[@]}"; do
-    model=$(parted /dev/$disk print | grep Model)
-    sd_disks+=("$disk")
+    model=$(parted -s /dev/$disk print | grep Model)
+    if [[ ! $model == *"nvme"* ]]; then
+        sd_disks+=("$disk")
+    fi
 done
 
 if [ "${sd_disks[0]}" == "" ]; then
@@ -71,16 +71,17 @@ fi
 echo "Available disk(s):"
 counter=0
 for disk in "${sd_disks[@]}"; do
-    model=$(parted /dev/$disk print | grep "Model" | cut -d " " -f2- )
-    size=$(parted /dev/$disk print | grep "Disk /dev/" | awk '{print $3}')
+    model=$(parted -s /dev/$disk print | grep "Model" | cut -d " " -f2- )
+    size=$(parted -s /dev/$disk print | grep "Disk /dev/" | awk '{print $3}')
     echo "$counter: $model /dev/$disk ($size)"
     counter=$(($counter + 1))
 done
 echo "Q: quit"
 
 read -p "Select a disk: " disk_choice
+disk_choice=${disk_choice,,}
 
-if [ "${disk_choice,,}" == "q" ]; then
+if [ "$disk_choice" == "q" ]; then
     cleanup_environment    
     echo "Script ended by user."
     exit
@@ -94,7 +95,7 @@ if [ "$disk_choice" == "" ] || [ "${sd_disks[disk_choice]}" == "" ]; then
 fi
 
 chosen_disk=${sd_disks[disk_choice]}
-size=$(parted /dev/$chosen_disk print | grep "Disk /dev/" | awk '{print $3}')
+size=$(parted -s /dev/$chosen_disk print | grep "Disk /dev/" | awk '{print $3}')
 echo "You have chosen: $chosen_disk ($size)"
 
 # Variable $size returns a string like 32.2GB (or MB or KB)
@@ -110,9 +111,9 @@ if [ "$size_id" == "GB" ]; then
     size_kb=$(($size_val * 1024 * 1024))
 fi
 
-minsize_kb=$((1024 * 1024 * 7))  # approximately 8GB in KB
+minsize_kb=$((1024 * 1024 * 14))  # approximately 16GB in KB
 if [ $size_kb -lt $minsize_kb ]; then
-    echo "Error: SD-card has insufficient capacity. Minimum size is 8GB."
+    echo "Error: SD-card has insufficient capacity. Minimum size is 16GB."
     cleanup_environment    
     echo "Script ended."
     exit
@@ -124,8 +125,9 @@ echo "D: Development (hostname = rpmsdev)"
 echo "Q: quit"
 
 read -p "Select a type: " type_choice
+type_choice=${type_choice,,}
 
-if [ "${type_choice,,}" == "q" ]; then
+if [ "$type_choice" == "q" ]; then
     cleanup_environment    
     echo "Script ended by user."
     exit
@@ -137,13 +139,60 @@ if [ "$type_choice" == "" ]; then
     echo "Script ended."
     exit
 fi
-if [ ${type_choice,,} != "p" ] && [ ${type_choice,,} != "d" ]; then
+if [ $type_choice != "p" ] && [ $type_choice != "d" ]; then
     echo "No type selected."
     cleanup_environment    
     echo "Script ended."
     exit
 fi
-echo "You have chosen: $type_choice $([ ${type_choice,,} == "p" ] && echo "=> production" || echo "=> development")"
+echo "You have chosen: $type_choice $([ $type_choice == "p" ] && echo "=> production" || echo "=> development")"
+
+echo "Language:"
+echo "E: English"
+echo "D: Dutch"
+echo "G: German"
+echo "F: French"
+echo "Q: quit"
+
+read -p "Select a language: " lang_choice
+lang_choice=${lang_choice,,}
+
+if [ "$lang_choice" == "q" ]; then
+    cleanup_environment    
+    echo "Script ended by user."
+    exit
+fi
+
+if [ "$lang_choice" == "" ]; then
+    echo "No language selected."
+    cleanup_environment    
+    echo "Script ended."
+    exit
+fi
+
+if [ $lang_choice != "e" ] && [ $lang_choice != "d" ] && [ $lang_choice != "g" ] && [ $lang_choice != "f" ]; then
+    echo "No language selected."
+    cleanup_environment    
+    echo "Script ended."
+    exit
+fi
+
+lang_choice_desc=""
+case $lang_choice in
+    e)
+        lang_choice_desc="English"
+        ;;
+    d)
+        lang_choice_desc="Dutch"
+        ;;
+    g)
+        lang_choice_desc="German"
+        ;;
+    f)
+        lang_choice_desc="French"
+        ;;
+esac
+echo "You have chosen: $lang_choice => $lang_choice_desc"
 
 read -r -p "Do you want to continue burning on $chosen_disk? [yes/NO] " start_install
 if [ "$start_install" != "yes" ]; then
@@ -157,10 +206,10 @@ if [ ! $(dpkg --list | grep wget | awk '{print $1}' | grep ii) ]; then
 fi
 echo "Downloading image..."
 wget -c --show-progress -P $working_dir -O $working_dir/$archive $image
-echo " => download complete"
+echo "... download complete."
 
 if [ "$(sha256sum $working_dir/$archive | cut -d' ' -f1)" != "$image_hash" ]; then
-    echo "Checksum of the downloaded image $archive failed."
+    echo "Checksum of the downloaded image $archive failed"
     cleanup_environment
     echo "Script ended with failure."
     exit
@@ -176,51 +225,61 @@ if [ -z $extracted_img ]; then
     echo "Script ended with failure."
     exit
 fi
-echo " => done extracting $working_dir/$archive"
+echo "... done extracting $working_dir/$archive."
 
 echo "Unmounting /dev/$chosen_disk partitions..."
 partitions=$(lsblk -l -n -p -e7 /dev/$chosen_disk | grep part | awk '{print $1}')
 for partition in $partitions; do
     sleep 3
-    umount -f "$partition"
-	if [ -n "$(df | grep $partition)" ]; then
-        echo "Failed to umount $partition."
-        cleanup_environment
-        echo "Script ended with failure."
-        exit
+    if [ -n "$(mount | grep "$partitions")" ]; then
+        umount -f "$partition"
+        if [ -n "$(df | grep $partition)" ]; then
+            echo "Failed to umount $partition."
+            cleanup_environment
+            echo "Script ended with failure."
+            exit
+        fi
+        echo "- partition $partition unmounted."
     fi
-   	echo "Partition $partition successfully unmounted."
 done
-hdparm -z /dev/$chosen_disk
-echo " => done unmounting /dev/$chosen_disk partitions"
+hdparm -z /dev/$chosen_disk > /dev/null
+echo "... done unmounting /dev/$chosen_disk partitions."
 
 echo "Start wiping $chosen_disk..."
 wipefs -a "/dev/$chosen_disk"
-hdparm -z /dev/$chosen_disk
-echo " => done wiping $chosen_disk."
+hdparm -z /dev/$chosen_disk > /dev/null
+echo "... done wiping $chosen_disk."
 
 if [ ! $(dpkg --list | grep gddrescue | awk '{print $1}' | grep ii) ]; then 
     apt install gddrescue -y
 fi
 echo "Start burning $extracted_img to $chosen_disk..."
 ddrescue -D --force $extracted_img "/dev/$chosen_disk"
-hdparm -z /dev/$chosen_disk
+hdparm -z /dev/$chosen_disk > /dev/null
 sleep 3  # Give the OS some time to reread
-echo " => done burning $chosen_disk."
+echo "... done burning $chosen_disk."
 
 if [ ! -d /dev/disk/by-label ]; then
-	echo "/dev/disk/by-label doesn't exist"
+	echo "/dev/disk/by-label doesn't exist."
     cleanup_environment
     echo "Script ended with failure."
     exit
 fi
 
-echo "Activate SSH..."
+echo "Mount boot"
 partition=$(ls -l /dev/disk/by-label | grep "boot" | grep -oE "$chosen_disk.*$")
 mount_partition
+
+echo "Activate SSH..."
 touch $mount_point/ssh
+echo "... SSH has been activated on $chosen_disk."
+
+echo "Unmount boot"
 unmount_partition
-echo " => SSH has been activated on $chosen_disk."
+
+echo "Mount rootfs"
+partition=$(ls -l /dev/disk/by-label | grep "rootfs" | grep -oE "$chosen_disk.*$")
+mount_partition
 
 if [ ${type_choice,,} == "p" ]; then
     hostname="rpms"
@@ -228,12 +287,16 @@ else
     hostname="rpmsdev"
 fi
 echo "Change hostname to $hostname..."
-partition=$(ls -l /dev/disk/by-label | grep "rootfs" | grep -oE "$chosen_disk.*$")
-mount_partition
 sed -i -e "s/raspberrypi/$hostname/g" $mount_point/etc/hostname
 sed -i -e "s/raspberrypi/$hostname/g" $mount_point/etc/hosts
-unmount_partition
-echo " => done changing hostname."
+echo "... done changing hostname."
+
+echo "Set language..."
+echo $lang_choice > $mount_point/etc/lang-choice.txt
+echo "... language has been set to $lang_choice."
+
+echo "Unmount rootfs"
+unmount_partition  
 
 cleanup_environment
 echo "Script ended successfully."
